update tests for watchlist
