+++ frontend-opal/tests/WatchListTest/WatchListItem.test.js
@@ -6,50 +6,15 @@ import thunk from 'redux-thunk';
 
 import reducers from 'reducers/root';
 import WatchListItem from 'components/WatchList/WatchListItem/WatchListItem';
+import presetProps from 'components/WatchList/presetProps';
 
 describe('<WatchListItem />', () => {
   it('renders correctly', () => {
-    const info = [
-      {
-        name: 'MSFT',
-        num: '141.21',
-        increase: '+0.23',
-        whole: '1.2M'
-      },
-      {
-        name: 'AMZN',
-        num: '2108.2',
-        increase: '-0.06',
-        whole: '609K'
-      },
-      {
-        name: 'AAPL',
-        num: '250.7',
-        increase: '+1.63',
-        whole: '810K'
-      }
-    ];
-    const chartData = [
-      { x: 1, y: 8 },
-      { x: 2, y: 7 },
-      { x: 3, y: 5 }
-    ];
-    const stockData = [
-      { name: '52 wk low', value: '231.6' },
-      { name: '52 wk high', value: '250.7' },
-      { name: 'P/E', value: '20.3' },
-      { name: 'EPS', value: '1.2' },
-      { name: 'DivYield', value: '2.34%' }
-    ];
-    const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
-    const store = createStore(reducers, composeEnhancers(applyMiddleware(thunk)));
-    const tree = renderer
-      .create(
-        <Provider store={store}>
-          <WatchListItem info={info} stockData={stockData} chartData={chartData} />
-        </Provider>
-      )
-      .toJSON();
+    const info = {
+      item: '',
+      data: [presetProps.defaultQuote]
+    };
+    const tree = renderer.create(<WatchListItem item={info.item} data={info.data} />).toJSON();
     expect(tree).toMatchSnapshot();
   });
 });
